requirements.txt
asgiref==3.8.1
boto3==1.37.10
botocore==1.37.10
Django==4.2.18
django-cors-headers==4.6.0
django-rest-passwordreset==1.5.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg==3.2.4
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
